Add files via upload
implement a system that successfully evolves solutions to the two
minimisation fitness functions below (separately) and demonstrate the effects of parameter changes,
through graphs and tables, including your understanding of what is happening.
Approaches you might like to consider are different forms of mutation and crossover, as well as selection.
For the final section it is acceptable to use implementations of other related algorithms for comparison with
your own code on the functions here and other benchmark functions. E.g., the use of tools such as Matlab
for non-evolutionary algorithms is acceptable.